Disable SSL (allow only TLS) and fixed a bug with loopback interface detection (NPE if network interface name = "lo0")
