+++ udc/src/main/java/io/crate/udc/ping/PingTask.java
@@ -35,10 +35,10 @@ import org.elasticsearch.common.xcontent.XContentFactory;
 import org.elasticsearch.http.HttpServerTransport;
 import org.hyperic.sigar.OperatingSystem;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
+import javax.net.ssl.*;
+import java.io.*;
 import java.net.*;
+import java.security.KeyStore;
 import java.security.NoSuchAlgorithmException;
 import java.util.*;
 import java.util.concurrent.ExecutionException;
@@ -55,6 +55,7 @@ public class PingTask extends TimerTask {
     private final ClusterIdService clusterIdService;
     private final HttpServerTransport httpServerTransport;
     private final String pingUrl;
+    private SSLSocketFactory ssf;
 
     private AtomicLong successCounter = new AtomicLong(0);
     private AtomicLong failCounter = new AtomicLong(0);
@@ -67,6 +68,12 @@ public class PingTask extends TimerTask {
         this.clusterIdService = clusterIdService;
         this.httpServerTransport = httpServerTransport;
         this.pingUrl = pingUrl;
+
+        try {
+            initializeTrustStore();
+        } catch (Exception e) {
+            logger.error("Failed initializing TrustStore: {}", e.getMessage());
+        }
     }
 
     @SuppressWarnings("unchecked")
@@ -108,17 +115,19 @@ public class PingTask extends TimerTask {
         try {
             NetworkInterface networkInterface = NetworkInterface.getByInetAddress(inetAddress);
             if (networkInterface != null) {
-                if (networkInterface.getName().equals("lo")) {
+                if (networkInterface.getName().startsWith("lo")) {
                     hardwareAddress = "loopback device";
                 } else {
                     byte[] hardwareAddressBytes = networkInterface.getHardwareAddress();
-                    StringBuilder sb = new StringBuilder(18);
-                    for (byte b : hardwareAddressBytes) {
-                        if (sb.length() > 0)
-                            sb.append(':');
-                        sb.append(String.format("%02x", b));
+                    if(hardwareAddressBytes != null) {
+                        StringBuilder sb = new StringBuilder(18);
+                        for (byte b : hardwareAddressBytes) {
+                            if (sb.length() > 0)
+                                sb.append(':');
+                            sb.append(String.format("%02x", b));
+                        }
+                        hardwareAddress = sb.toString();
                     }
-                    hardwareAddress = sb.toString();
                 }
             }
 
@@ -177,7 +186,15 @@ public class PingTask extends TimerTask {
             if (logger.isDebugEnabled()) {
                 logger.debug("Sending UDC information to {}...", url);
             }
-            HttpURLConnection conn = (HttpURLConnection)url.openConnection();
+            HttpURLConnection conn = null;
+            if(url.getProtocol().equalsIgnoreCase("https")) {
+                HttpsURLConnection sslConn = (HttpsURLConnection)url.openConnection();
+                sslConn.setSSLSocketFactory(ssf);
+                //TODO consider hostname verification
+                conn = sslConn;
+            } else {
+                conn = (HttpURLConnection)url.openConnection();
+            }
             conn.setConnectTimeout((int)HTTP_TIMEOUT.millis());
             conn.setReadTimeout((int)HTTP_TIMEOUT.millis());
 
@@ -202,5 +219,40 @@ public class PingTask extends TimerTask {
             }
             failCounter.incrementAndGet();
         }
+    }
+
+    /*
+     * Initialize our own TrustStore including StartCom CA
+     *
+     * TrustStore was generated by:
+     *  keytool -import -file sub.class2.server.ca.pem -alias CAAlias -keystore truststore.ts -storepass udc.crate
+     *
+     */
+    private void initializeTrustStore() throws Exception {
+        String trustStorePath = "/ssl/truststore.ts";
+        String trustPassword = "udc.crate";
+
+        // load our key store as a stream and initialize a KeyStore
+        InputStream trustStream = this.getClass().getResourceAsStream(trustStorePath);
+        if (trustStream == null) {
+            throw new FileNotFoundException("Resource [" + trustStorePath + "] not found in classpath");
+        }
+        KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType());
+
+        // load the stream to our store
+        trustStore.load(trustStream, trustPassword.toCharArray());
+
+        // initialize a trust manager factory with the trusted store
+        TrustManagerFactory trustFactory =
+                TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+        trustFactory.init(trustStore);
+
+        // get the trust managers from the factory
+        TrustManager[] trustManagers = trustFactory.getTrustManagers();
+
+        // initialize an ssl context to use these managers and set as default
+        SSLContext sslContext = SSLContext.getInstance("TLS");
+        sslContext.init(null, trustManagers, null);
+        ssf = sslContext.getSocketFactory();
     }
 }
