+++ udc/src/main/java/io/crate/udc/plugin/UDCPlugin.java
@@ -29,12 +29,6 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.plugins.AbstractPlugin;
 
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.TrustManagerFactory;
-import java.io.FileNotFoundException;
-import java.io.InputStream;
-import java.security.KeyStore;
 import java.util.Collection;
 import java.util.concurrent.TimeUnit;
 
@@ -60,12 +54,6 @@ public class UDCPlugin extends AbstractPlugin {
 
     public UDCPlugin(Settings settings) {
         this.settings = settings;
-
-        try {
-            initializeTrustStore();
-        } catch (Exception e) {
-            logger.error("Failed initializing TrustStore: {}", e.getMessage());
-        }
     }
 
     @Override
@@ -87,40 +75,5 @@ public class UDCPlugin extends AbstractPlugin {
     @Override
     public String description() {
         return "crate plugin for Usage Data Collection (UDC)";
-    }
-
-    /*
-     * Initialize our own TrustStore including StartCom CA
-     *
-     * TrustStore was generated by:
-     *  keytool -import -file sub.class2.server.ca.pem -alias CAAlias -keystore truststore.ts -storepass udc.crate
-     *
-     */
-    private void initializeTrustStore() throws Exception {
-        String trustStorePath = "/ssl/truststore.ts";
-        String trustPassword = "udc.crate";
-
-        // load our key store as a stream and initialize a KeyStore
-        InputStream trustStream = this.getClass().getResourceAsStream(trustStorePath);
-        if (trustStream == null) {
-            throw new FileNotFoundException("Resource [" + trustStorePath + "] not found in classpath");
-        }
-        KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType());
-
-        // load the stream to our store
-        trustStore.load(trustStream, trustPassword.toCharArray());
-
-        // initialize a trust manager factory with the trusted store
-        TrustManagerFactory trustFactory =
-                TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-        trustFactory.init(trustStore);
-
-        // get the trust managers from the factory
-        TrustManager[] trustManagers = trustFactory.getTrustManagers();
-
-        // initialize an ssl context to use these managers and set as default
-        SSLContext sslContext = SSLContext.getInstance("SSL");
-        sslContext.init(null, trustManagers, null);
-        SSLContext.setDefault(sslContext);
     }
 }
